reorganizing structure for pip install
:heavy_check_mark: move testing up to one level, sub folders aren't playing nice with poetry, typer combo
:heavy_check_mark: remove src directory, and have one app directory named postporcessing with a main.py
:heavy_check_mark: re-org fixes issue with pip install
