+++ postprocessing_variant_calls/main.py
@@ -1,6 +1,6 @@
 # Main Import 
 import typer
-import vardict.vardict_process
+from .vardict import vardict_process
 import logging
 import time
 # setup logger 
@@ -8,7 +8,7 @@ logger = logging.getLogger("filter")
 app = typer.Typer()
 
 # Vardict filter 
-app.add_typer(vardict.vardict_process.app, name="vardict")
+app.add_typer(vardict_process.app, name="vardict")
 # Haplo filter 
 #### Insert Here ####
